requirements.txt
fastapi
uvicorn
sqlalchemy>=2
asyncpg
jinja2
python-multipart
python-dotenv
pydantic-settings
pytz
starlette